# Specs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Blog API",
    "description": "A Clean Architecture API for Blogging Platform",
    "contact": {
      "name": "Developer",
      "email": "dev@example.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/Orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/Orders/my": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Gets the current authenticated user's orders with pagination and status filtering.",
        "parameters": [
          {
            "name": "PageNumber",
            "in": "query",
            "description": "The page number to retrieve, defaults to 1.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "description": "The page size to retrieve, defaults to 10.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "OrderStatus",
            "in": "query",
            "description": "Optional status to filter the user's orders.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Orders/{id}/details": {
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "Updates order shipping details and/or payment method (Combined flow).",
        "description": "LUẬT NGHIỆP VỤ: \r\n1. Chỉ cho phép thực thi khi đơn hàng ở trạng thái 'Pending' và trạng thái thanh toán 'Pending'. Nếu không, trả về lỗi 400 Bad Request.\r\n2. Cho phép cập nhật thông tin vận chuyển và phương thức thanh toán. \r\n3. Nếu phương thức thanh toán thay đổi so với phương thức cũ, hệ thống tự động xóa PaymentUrl cũ (set null) để đảm bảo tính nhất quán của giao dịch.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDetailsCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDetailsCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDetailsCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Orders/{id}/pay": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Invokes the online payment processing to retrieve a fresh payment URL (Lazy generation).",
        "description": "LUẬT NGHIỆP VỤ: \r\n1. Chỉ cho phép sinh link thanh toán cho các đơn hàng có trạng thái 'Pending' và trạng thái thanh toán 'Pending'.\r\n2. Chỉ hỗ trợ các phương thức thanh toán trực tuyến (VNPAY, MOMO, PAYOS). Không hỗ trợ COD.\r\n3. Link thanh toán sẽ được tự động cấu hình callback/webhook động tương ứng theo cổng được chọn để cập nhật tự động trạng thái đơn hàng.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Orders/{orderId}/items": {
      "post": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddOrderItemCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddOrderItemCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddOrderItemCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Orders/{orderId}/items/{orderItemId}": {
      "delete": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "orderItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Orders/statistics": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Gets order statistics for administrators.",
        "description": "LUẬT NGHIỆP VỤ & TỐI ƯU HÓA: \r\n1. API thống kê doanh thu, tổng số đơn hàng, giá trị đơn hàng trung bình, phân nhóm theo trạng thái và theo tháng cho quản trị viên.\r\n2. Được cấu hình để thực hiện tối ưu hóa truy vấn thông qua cơ chế Projection, tránh tải các thực thể (entities) đầy đủ lên bộ nhớ để ngăn ngừa lỗi Over-fetching và tràn RAM khi khối lượng đơn hàng lớn.",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "description": "Optional start date to filter orders.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "Optional end date to filter orders.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddOrderItemCommand": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "Address": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateOrderCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orderStatus": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Command to update order status."
      },
      "UpdateOrderDetailsCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "shippingAddress": {
            "$ref": "#/components/schemas/Address"
          },
          "paymentMethod": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Command to update general order details (Shipping address, receiver, payment method)."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter JWT Bearer token **_only_**",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}